refactor(app): extract message interactions
