+++ src/app.rs
@@ -32,6 +32,7 @@ pub mod media_support;
 pub mod message_action_diagnostics;
 pub mod message_actions;
 pub mod message_ingestion;
+pub mod message_interactions;
 pub mod message_navigation;
 pub mod message_opening;
 pub mod notifications;
